rename file with component name
